initial checkin
The sole purpose of this silly game is to entertain my family.  I put it
together in haste (so the code is not up to my standards) and there are
a lot of copyrighted content in the repo that I probably had no right to
redistribute.  My only excuse is that I mean no harm, and may take it down
altogether.